feat(navbar): redesign of navbar
