+++ craco.config.js
@@ -40,6 +40,19 @@ module.exports = {
     ]
   },
   webpack: {
+    configure: {
+      module: {
+        rules: [
+          {
+            type: 'javascript/auto',
+
+            test: /\.mjs$/,
+
+            include: /node_modules/
+          }
+        ]
+      }
+    },
     plugins: [
       new CopyPlugin({
         patterns: [
