+++ package.json
@@ -3,7 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "engines": {
-    "node": ">=16.18.1 <=20.8.1"
+    "node": ">=16.18.1 <=20.11.4"
   },
   "dependencies": {
     "@craco/craco": "^6.1.1",
@@ -11,6 +11,8 @@
     "@fvictorio/newton-raphson-method": "^1.0.5",
     "@material-ui/core": "^4.12.3",
     "@material-ui/pickers": "^3.3.10",
+    "@radix-ui/react-popover": "^1.0.7",
+    "@react-spring/web": "^9.2.4",
     "@reality.eth/reality-eth-lib": "^2.0.13",
     "@reduxjs/toolkit": "^1.9.7",
     "@sentry/cli": "^2.21.5",
