+++ src/assets/icons/PolkamarketsLogo.tsx
@@ -3,49 +3,65 @@ import React from 'react';
 function PolkamarketsLogo() {
   return (
     <svg
+      width="167"
+      height="30"
+      viewBox="0 0 167 30"
+      fill="none"
       xmlns="http://www.w3.org/2000/svg"
-      height="20"
-      style={{ transform: 'translateY(-7%)' }}
-      viewBox="0 0 142 24"
     >
+      <path
+        fillRule="evenodd"
+        clipRule="evenodd"
+        d="M35.84 17.7406V22.2206H33V7.52063H35.44H35.84H38.04C39.2667 7.52063 40.3067 7.72063 41.16 8.12063C42.0267 8.52063 42.6867 9.09396 43.14 9.84063C43.5933 10.574 43.82 11.4473 43.82 12.4606V12.7806C43.82 13.794 43.5933 14.674 43.14 15.4206C42.6867 16.154 42.0267 16.7273 41.16 17.1406C40.3067 17.5406 39.2667 17.7406 38.04 17.7406H35.84ZM35.84 15.2806H38.28C38.8533 15.2806 39.3333 15.1673 39.72 14.9406C40.12 14.714 40.42 14.4006 40.62 14.0006C40.8333 13.6006 40.94 13.1406 40.94 12.6206C40.94 12.1006 40.8333 11.6406 40.62 11.2406C40.42 10.8406 40.12 10.534 39.72 10.3206C39.3333 10.094 38.8533 9.98063 38.28 9.98063H35.84V15.2806ZM58.702 9.70063V22.2206H61.482V9.70063V7.62063H58.702H57.342V9.70063H58.702ZM71.8558 22.2206L68.2158 17.1406H67.0358V22.2206H64.2558V7.62063H67.0358V16.5902L71.2158 11.3606H74.1158L70.4223 15.9905L74.9958 22.2206H71.8558ZM129.81 17.1406L133.45 22.2206H136.59L132.016 15.9905L135.71 11.3606H132.81L128.63 16.5902V7.62063H125.85V22.2206H128.63V17.1406H129.81ZM151.684 21.9806C152.337 22.234 153.164 22.3606 154.164 22.3606H155.724V19.9806H154.024C153.491 19.9806 153.077 19.8406 152.784 19.5606C152.504 19.2673 152.364 18.854 152.364 18.3206L152.374 13.4006H155.724V11.3606H152.378L152.384 8.30063H149.784L149.778 11.3606H148.044V13.4006H149.774L149.764 18.1406C149.764 19.194 149.924 20.034 150.244 20.6606C150.564 21.274 151.044 21.714 151.684 21.9806ZM157.999 21.6206C158.853 22.2473 160.026 22.5606 161.519 22.5606C162.479 22.5606 163.313 22.414 164.019 22.1206C164.726 21.8273 165.273 21.4206 165.659 20.9006C166.059 20.3673 166.259 19.7406 166.259 19.0206C166.259 18.0473 165.893 17.274 165.159 16.7006C164.439 16.1273 163.439 15.774 162.159 15.6406L161.199 15.5406C160.613 15.474 160.186 15.3473 159.919 15.1606C159.666 14.974 159.539 14.7073 159.539 14.3606C159.539 14.0006 159.699 13.7073 160.019 13.4806C160.339 13.254 160.793 13.1406 161.379 13.1406C162.033 13.1406 162.533 13.274 162.879 13.5406C163.226 13.8073 163.413 14.1273 163.439 14.5006H165.939C165.899 13.3673 165.466 12.5073 164.639 11.9206C163.813 11.334 162.733 11.0406 161.399 11.0406C160.546 11.0406 159.786 11.174 159.119 11.4406C158.466 11.694 157.953 12.074 157.579 12.5806C157.206 13.0873 157.019 13.714 157.019 14.4606C157.019 15.394 157.346 16.1406 157.999 16.7006C158.653 17.2606 159.586 17.6073 160.799 17.7406L161.759 17.8406C162.479 17.9073 162.986 18.0473 163.279 18.2606C163.586 18.474 163.739 18.7606 163.739 19.1206C163.739 19.534 163.539 19.8606 163.139 20.1006C162.753 20.3406 162.226 20.4606 161.559 20.4606C160.759 20.4606 160.173 20.3073 159.799 20.0006C159.426 19.694 159.219 19.354 159.179 18.9806H156.679C156.719 20.1006 157.159 20.9806 157.999 21.6206ZM142.302 22.6006C141.368 22.6006 140.548 22.4406 139.842 22.1206C139.148 21.8006 138.568 21.374 138.102 20.8406C137.648 20.294 137.302 19.6873 137.062 19.0206C136.835 18.354 136.722 17.674 136.722 16.9806V16.6006C136.722 15.8806 136.835 15.1873 137.062 14.5206C137.302 13.8406 137.648 13.2406 138.102 12.7206C138.568 12.1873 139.142 11.7673 139.822 11.4606C140.502 11.1406 141.288 10.9806 142.182 10.9806C143.355 10.9806 144.335 11.2406 145.122 11.7606C145.922 12.2673 146.522 12.9406 146.922 13.7806C147.322 14.6073 147.522 15.5006 147.522 16.4606V17.4606H139.358C139.405 17.8797 139.5 18.2663 139.642 18.6206C139.855 19.1406 140.182 19.554 140.622 19.8606C141.062 20.154 141.622 20.3006 142.302 20.3006C142.982 20.3006 143.535 20.1673 143.962 19.9006C144.388 19.6206 144.662 19.2806 144.782 18.8806H147.342C147.182 19.6273 146.875 20.2806 146.422 20.8406C145.968 21.4006 145.388 21.834 144.682 22.1406C143.988 22.4473 143.195 22.6006 142.302 22.6006ZM139.642 14.9206C139.537 15.1751 139.458 15.4551 139.405 15.7606H144.866C144.817 15.4149 144.736 15.1015 144.622 14.8206C144.422 14.3273 144.115 13.9473 143.702 13.6806C143.302 13.414 142.795 13.2806 142.182 13.2806C141.568 13.2806 141.048 13.4206 140.622 13.7006C140.195 13.9806 139.868 14.3873 139.642 14.9206ZM117.91 22.2206V11.3606H120.11V14.8827C120.233 13.8454 120.546 13.0247 121.05 12.4206C121.716 11.6206 122.696 11.2206 123.99 11.2206H124.39V13.6406H123.63C122.696 13.6406 121.97 13.894 121.45 14.4006C120.943 14.894 120.69 15.614 120.69 16.5606V22.2206H117.91ZM113.132 20.1292V22.2206H115.332V15.4806C115.332 14.4673 115.139 13.6473 114.752 13.0206C114.379 12.394 113.819 11.934 113.072 11.6406C112.339 11.3473 111.412 11.2006 110.292 11.2006C109.932 11.2006 109.545 11.2073 109.132 11.2206C108.732 11.2206 108.339 11.234 107.952 11.2606C107.565 11.2873 107.219 11.314 106.912 11.3406V13.7006C107.285 13.674 107.705 13.654 108.172 13.6406C108.652 13.614 109.125 13.594 109.592 13.5806C110.059 13.5673 110.459 13.5606 110.792 13.5606C111.432 13.5606 111.905 13.714 112.212 14.0206C112.519 14.3273 112.672 14.794 112.672 15.4206V15.5206H110.452C109.492 15.5206 108.652 15.6473 107.932 15.9006C107.225 16.154 106.672 16.534 106.272 17.0406C105.885 17.5473 105.692 18.2006 105.692 19.0006C105.692 19.734 105.852 20.3606 106.172 20.8806C106.505 21.4006 106.965 21.8006 107.552 22.0806C108.152 22.3606 108.845 22.5006 109.632 22.5006C110.392 22.5006 111.032 22.3673 111.552 22.1006C112.085 21.8206 112.499 21.4206 112.792 20.9006C112.929 20.6676 113.043 20.4104 113.132 20.1292ZM112.672 17.1806H110.412C109.772 17.1806 109.279 17.3406 108.932 17.6606C108.599 17.9673 108.432 18.3673 108.432 18.8606C108.432 19.354 108.599 19.754 108.932 20.0606C109.279 20.3673 109.772 20.5206 110.412 20.5206C110.799 20.5206 111.152 20.454 111.472 20.3206C111.805 20.174 112.079 19.934 112.292 19.6006C112.519 19.254 112.645 18.7873 112.672 18.2006V17.1806ZM87.4011 22.2206V11.3606H89.6011V14.0144C89.661 13.7525 89.7343 13.5079 89.8211 13.2806C90.1011 12.534 90.5144 11.974 91.0611 11.6006C91.6211 11.214 92.3144 11.0206 93.1411 11.0206H93.2611C94.1011 11.0206 94.7944 11.214 95.3411 11.6006C95.9011 11.974 96.3144 12.534 96.5811 13.2806C96.6055 13.3445 96.6288 13.4098 96.6511 13.4764C96.6734 13.4098 96.6967 13.3445 96.7211 13.2806C97.0144 12.534 97.4344 11.974 97.9811 11.6006C98.5411 11.214 99.2344 11.0206 100.061 11.0206H100.181C101.021 11.0206 101.721 11.214 102.281 11.6006C102.841 11.974 103.261 12.534 103.541 13.2806C103.834 14.014 103.981 14.9273 103.981 16.0206V22.2206H101.201V15.7606C101.201 15.0806 101.028 14.5406 100.681 14.1406C100.334 13.7273 99.8411 13.5206 99.2011 13.5206C98.5611 13.5206 98.0478 13.734 97.6611 14.1606C97.2744 14.574 97.0811 15.134 97.0811 15.8406V22.2206H94.3011V15.7606C94.3011 15.0806 94.1278 14.5406 93.7811 14.1406C93.4344 13.7273 92.9411 13.5206 92.3011 13.5206C91.6611 13.5206 91.1478 13.734 90.7611 14.1606C90.3744 14.574 90.1811 15.134 90.1811 15.8406V22.2206H87.4011ZM82.6233 20.1292V22.2206H84.8233V15.4806C84.8233 14.4673 84.6299 13.6473 84.2433 13.0206C83.8699 12.394 83.31 11.934 82.5633 11.6406C81.83 11.3473 80.9033 11.2006 79.7833 11.2006C79.4233 11.2006 79.0366 11.2073 78.6233 11.2206C78.2233 11.2206 77.8299 11.234 77.4433 11.2606C77.0566 11.2873 76.7099 11.314 76.4033 11.3406V13.7006C76.7766 13.674 77.1966 13.654 77.6633 13.6406C78.1433 13.614 78.6166 13.594 79.0833 13.5806C79.5499 13.5673 79.95 13.5606 80.2833 13.5606C80.9233 13.5606 81.3966 13.714 81.7033 14.0206C82.01 14.3273 82.1633 14.794 82.1633 15.4206V15.5206H79.9433C78.9833 15.5206 78.1433 15.6473 77.4233 15.9006C76.7166 16.154 76.1633 16.534 75.7633 17.0406C75.3766 17.5473 75.1833 18.2006 75.1833 19.0006C75.1833 19.734 75.3433 20.3606 75.6633 20.8806C75.9966 21.4006 76.4566 21.8006 77.0433 22.0806C77.6433 22.3606 78.3366 22.5006 79.1233 22.5006C79.8833 22.5006 80.5233 22.3673 81.0433 22.1006C81.5766 21.8206 81.9899 21.4206 82.2833 20.9006C82.4207 20.6676 82.534 20.4104 82.6233 20.1292ZM82.1633 17.1806H79.9033C79.2633 17.1806 78.77 17.3406 78.4233 17.6606C78.09 17.9673 77.9233 18.3673 77.9233 18.8606C77.9233 19.354 78.09 19.754 78.4233 20.0606C78.77 20.3673 79.2633 20.5206 79.9033 20.5206C80.2899 20.5206 80.6433 20.454 80.9633 20.3206C81.2966 20.174 81.57 19.934 81.7833 19.6006C82.0099 19.254 82.1366 18.7873 82.1633 18.2006V17.1806ZM50.6078 22.6006C49.6478 22.6006 48.8011 22.4473 48.0678 22.1406C47.3345 21.834 46.7145 21.4206 46.2078 20.9006C45.7011 20.3673 45.3145 19.7673 45.0478 19.1006C44.7945 18.434 44.6678 17.7406 44.6678 17.0206V16.6006C44.6678 15.854 44.8011 15.1473 45.0678 14.4806C45.3478 13.8006 45.7411 13.2006 46.2478 12.6806C46.7678 12.1473 47.3945 11.734 48.1278 11.4406C48.8611 11.134 49.6878 10.9806 50.6078 10.9806C51.5278 10.9806 52.3545 11.134 53.0878 11.4406C53.8211 11.734 54.4411 12.1473 54.9478 12.6806C55.4678 13.2006 55.8611 13.8006 56.1278 14.4806C56.3945 15.1473 56.5278 15.854 56.5278 16.6006V17.0206C56.5278 17.7406 56.3945 18.434 56.1278 19.1006C55.8745 19.7673 55.4945 20.3673 54.9878 20.9006C54.4811 21.4206 53.8611 21.834 53.1278 22.1406C52.3945 22.4473 51.5545 22.6006 50.6078 22.6006ZM50.6078 20.2206C51.2878 20.2206 51.8611 20.074 52.3278 19.7806C52.7945 19.474 53.1478 19.0673 53.3878 18.5606C53.6278 18.0406 53.7478 17.454 53.7478 16.8006C53.7478 16.134 53.6211 15.5473 53.3678 15.0406C53.1278 14.5206 52.7678 14.114 52.2878 13.8206C51.8211 13.514 51.2611 13.3606 50.6078 13.3606C49.9545 13.3606 49.3878 13.514 48.9078 13.8206C48.4411 14.114 48.0811 14.5206 47.8278 15.0406C47.5745 15.5473 47.4478 16.134 47.4478 16.8006C47.4478 17.454 47.5678 18.0406 47.8078 18.5606C48.0611 19.0673 48.4211 19.474 48.8878 19.7806C49.3545 20.074 49.9278 20.2206 50.6078 20.2206Z"
+        fill="white"
+      />
+      <path
+        d="M12.8467 0.5L0.900391 14.9036L12.8693 29.4655L24.9004 14.9318L12.8467 0.5Z"
+        fill="url(#paint0_linear_182_388)"
+      />
+      <path
+        d="M12.8693 29.4657L0.900391 14.9037L15.9364 11.2604L12.8693 29.4657Z"
+        fill="url(#paint1_linear_182_388)"
+      />
+      <path
+        d="M12.8467 0.5L24.9004 14.9318L15.9364 11.2546L12.8467 0.5Z"
+        fill="url(#paint2_linear_182_388)"
+      />
       <defs>
-        <style>
-          {`
-            .cls-1 {
-              fill: #3033be;
-            }
-
-            .cls-2 {
-              fill: #f2f0ff;
-            }
-
-            .theme--light .cls-2 {
-              fill: #3033be;
-            }
-
-            .cls-3 {
-              fill: #5751fc;
-            }
-
-            .cls-4 {
-              fill: #51a0fc;
-            }
-        `}
-        </style>
+        <linearGradient
+          id="paint0_linear_182_388"
+          x1="12.9004"
+          y1="0.5"
+          x2="12.9004"
+          y2="29.4655"
+          gradientUnits="userSpaceOnUse"
+        >
+          <stop stopColor="#8F5EE1" />
+          <stop offset="1" stopColor="#5B3AAB" />
+        </linearGradient>
+        <linearGradient
+          id="paint1_linear_182_388"
+          x1="8.4184"
+          y1="11.2604"
+          x2="8.4184"
+          y2="29.4657"
+          gradientUnits="userSpaceOnUse"
+        >
+          <stop stopColor="#21165A" />
+          <stop offset="1" stopColor="#322076" />
+        </linearGradient>
+        <linearGradient
+          id="paint2_linear_182_388"
+          x1="18.8736"
+          y1="14.9318"
+          x2="18.8736"
+          y2="1.40777"
+          gradientUnits="userSpaceOnUse"
+        >
+          <stop stopColor="#C03739" />
+          <stop offset="1" stopColor="#FD8E68" />
+        </linearGradient>
       </defs>
-      <path
-        className="cls-3"
-        d="M9.914,0L0,11.934l9.933,12.065,9.984-12.042L9.914,0Z"
-      />
-      <path
-        className="cls-1"
-        d="M9.933,24L0,11.935l12.478-3.019-2.545,15.084Z"
-      />
-      <path className="cls-4" d="M9.914,0l10.003,11.958-7.439-3.047L9.914,0Z" />
-      <path
-        className="cls-2"
-        d="M28.011,4.874h5.457c1.391,0,2.576,.478,3.554,1.424,.967,.957,1.446,2.12,1.446,3.511s-.478,2.554-1.446,3.511c-.978,.956-2.163,1.424-3.554,1.424h-2.576v4.837h-2.891V4.874h.011Zm2.891,7.163h2.576c.609,0,1.12-.207,1.522-.63,.402-.435,.609-.967,.609-1.598,0-.641-.207-1.174-.609-1.598-.402-.424-.913-.63-1.522-.63h-2.576v4.456Zm16.902,6.239c-1.054,1.065-2.37,1.598-3.935,1.598s-2.87-.533-3.935-1.598c-1.054-1.065-1.587-2.38-1.587-3.946,0-1.554,.533-2.87,1.587-3.935,1.076-1.076,2.381-1.62,3.935-1.62,1.544,0,2.859,.543,3.935,1.62,1.076,1.076,1.609,2.391,1.609,3.935,0,1.554-.533,2.87-1.609,3.946Zm-5.957-1.87c.543,.543,1.217,.815,2.033,.815,.804,0,1.489-.272,2.033-.815,.544-.543,.815-1.239,.815-2.076s-.272-1.533-.815-2.076-1.217-.815-2.033-.815c-.804,0-1.489,.272-2.033,.815-.533,.565-.794,1.25-.794,2.076s.261,1.522,.794,2.076Zm10.989,3.174h-2.696V4.244h2.696v15.337Zm6.728-5.315l4.326,5.315h-3.141l-3.804-4.772v4.772h-2.696V4.874h2.696V13.7l3.598-4.619h3.217l-4.196,5.185Zm11.826-3.946v-1.239h2.696v10.511h-2.696v-1.239c-.804,1.022-1.946,1.533-3.413,1.533-1.391,0-2.598-.533-3.598-1.598-.989-1.076-1.489-2.391-1.489-3.946,0-1.543,.5-2.848,1.489-3.935,1-1.076,2.207-1.62,3.598-1.62,1.467-.011,2.609,.5,3.413,1.533Zm-4.978,6.152c.543,.543,1.239,.815,2.065,.815,.837,0,1.522-.272,2.065-.815,.554-.565,.837-1.272,.837-2.141s-.283-1.576-.837-2.12c-.543-.565-1.239-.837-2.065-.837-.837,0-1.522,.283-2.065,.837-.543,.543-.815,1.25-.815,2.12s.272,1.587,.815,2.141Zm19.902-7.652c1.185,0,2.131,.38,2.848,1.13,.707,.761,1.065,1.783,1.065,3.065v6.554h-1.381v-6.543c0-.891-.228-1.598-.685-2.098-.456-.5-1.076-.761-1.859-.761-.902,0-1.62,.304-2.131,.902-.511,.598-.772,1.5-.772,2.685v5.826h-1.391v-6.554c0-.913-.217-1.609-.652-2.098-.435-.5-1.033-.761-1.794-.761-.88,0-1.609,.304-2.174,.924-.554,.619-.837,1.511-.837,2.674v5.826h-1.381V9.081h1.381v1.533c.75-1.196,1.815-1.783,3.196-1.783,1.522,0,2.576,.641,3.174,1.913,.761-1.283,1.891-1.924,3.391-1.924Zm14.532,2.315v-2.054h1.381v10.511h-1.381v-2.054c-.967,1.543-2.369,2.315-4.228,2.315-1.489,0-2.772-.533-3.826-1.598-1.043-1.065-1.565-2.37-1.565-3.913s.522-2.848,1.565-3.913c1.065-1.065,2.337-1.598,3.826-1.598,1.859-.011,3.272,.761,4.228,2.304Zm-7.043,6.152c.793,.793,1.772,1.196,2.924,1.196s2.131-.402,2.924-1.196c.793-.826,1.195-1.815,1.195-2.967,0-1.163-.402-2.141-1.195-2.946-.794-.815-1.772-1.217-2.924-1.217s-2.13,.402-2.924,1.217c-.794,.793-1.196,1.783-1.196,2.946s.391,2.141,1.196,2.967Zm11.13-8.206v1.739c.642-1.272,1.739-1.913,3.305-1.913v1.348c-1,0-1.805,.293-2.403,.88-.597,.576-.902,1.446-.902,2.609v5.848h-1.38V9.081h1.38Zm7,5.043l5.609,5.467h-1.837l-5.228-5.065v5.065h-1.381V4.874h1.381V13.722l4.978-4.641h1.88l-5.402,5.043Zm10.229-5.304c1.565,0,2.804,.554,3.739,1.663,.967,1.098,1.445,2.391,1.445,3.891,0,.065-.011,.283-.043,.652h-9.185c.141,1.054,.587,1.891,1.337,2.522,.75,.63,1.696,.946,2.826,.946,1.587,0,2.717-.598,3.37-1.783l1.217,.717c-.456,.761-1.098,1.348-1.902,1.783-.794,.424-1.696,.63-2.696,.63-1.641,0-2.989-.522-4.011-1.554-1.032-1.033-1.543-2.359-1.543-3.946s.511-2.891,1.522-3.935c1.01-1.054,2.326-1.587,3.924-1.587Zm0,1.348c-1.087,0-2,.326-2.74,.967-.728,.641-1.152,1.489-1.293,2.543h7.804c-.152-1.12-.576-1.978-1.271-2.587-.729-.62-1.566-.924-2.5-.924Zm11.576-1.087v1.337h-3.011v6.369c0,.576,.108,.989,.337,1.239,.228,.239,.554,.37,1,.38,.445,0,1-.011,1.674-.044v1.217c-1.489,.228-2.598,.12-3.326-.315-.707-.446-1.066-1.272-1.066-2.478v-6.369h-2.195v-1.348h2.195v-2.522l1.381-.424v2.946h3.011v.011Zm2.163,2.75c0,.75,.956,1.467,2.152,1.783,.402,.098,.674,.163,.793,.206,.131,.033,.37,.109,.728,.228,.381,.109,.631,.207,.772,.293,.152,.087,.359,.217,.631,.402,.261,.163,.456,.348,.565,.522,.109,.163,.217,.391,.315,.674,.098,.261,.141,.554,.141,.859,0,.891-.358,1.63-1.065,2.185-.685,.565-1.587,.837-2.717,.837-.978,0-1.815-.217-2.533-.652-.706-.446-1.217-1.033-1.511-1.761l1.174-.696c.196,.543,.544,.978,1.044,1.304,.511,.304,1.119,.467,1.815,.467,.706,0,1.293-.141,1.739-.424s.674-.696,.674-1.261c0-.815-.717-1.413-2.152-1.783-.402-.098-.674-.163-.794-.185-.13-.044-.38-.12-.75-.228-.358-.13-.619-.228-.771-.315-.142-.087-.337-.206-.609-.38-.261-.185-.457-.359-.565-.522-.109-.185-.218-.402-.316-.674-.097-.283-.141-.576-.141-.88,0-.859,.337-1.565,1-2.141,.696-.576,1.554-.859,2.576-.859,.837,0,1.576,.185,2.218,.565,.652,.38,1.141,.902,1.445,1.576l-1.152,.674c-.391-.978-1.228-1.467-2.511-1.467-.63,0-1.152,.152-1.565,.446-.424,.283-.63,.685-.63,1.207Z"
-      />
     </svg>
   );
 }
