+++ src/components/Header/Header.module.scss
@@ -1,5 +1,5 @@
 .root {
-  border-bottom: 1px solid var(--color-border);
+  border-bottom: 1px solid #252C3B;
   display: flex;
   grid-area: header;
 }
