+++ src/components/Header/HeaderActions/HeaderActions.module.scss
@@ -6,26 +6,26 @@
   bottom: 0px;
   left: 0px;
   transition: background-color 200ms ease-in-out;
-  background-color: var(--color-surface-primary);
-  z-index: 1;
-
+  background-color: #0A0B0E;
+  justify-content: flex-end;
+  gap: 6px;
+  width: 100%;
+  padding: 0 var(--size-12);
+  color: #FFF;
+  font-size: 14px;
+  font-style: normal;
+  font-weight: 500;
+  line-height: 20px;
   @media (min-width: 1024px) {
-    background-color: unset;
     transition: unset;
     margin-left: auto;
     border-top: unset;
     position: unset;
     width: unset;
+    padding: 0;
   }
 }
-.gutterActions {
-  gap: 12px;
-}
-.reverse {
-  @media (min-width: 1024px) {
-    flex-direction: row-reverse;
-  }
-}
+
 .actions {
   &Group {
     display: inherit;
@@ -44,4 +44,12 @@
 .transactions {
   padding: var(--size-12) !important;
   margin-left: auto;
+}
+
+.createButton {
+  color: #FFF;
+  font-size: 14px !important;
+  font-style: normal !important;
+  font-weight: 500 !important;
+  line-height: 20px !important;
 }
