+++ src/components/Header/HeaderNav/HeaderNav.tsx
@@ -3,30 +3,25 @@ import { NavLink, Link } from 'react-router-dom';
 
 import cn from 'classnames';
 import { pages, community, ui, features } from 'config';
-import getPathname from 'helpers/getPathname';
 import isEmpty from 'lodash/isEmpty';
 import { useTheme } from 'ui';
 
 import * as Logos from 'assets/icons';
-import { ReactComponent as V2Badge } from 'assets/icons/svgs/v2-badge.svg';
 
 import { Button } from 'components/Button';
 import CreateMarket from 'components/CreateMarket';
 import Feature from 'components/Feature';
-import HelpButton from 'components/HelpButton';
+import HelpButton from 'components/Header/HelpButton';
+import ProfileSignin from 'components/Header/ProfileSignin';
 import HowToPlayButton from 'components/HowToPlayButton';
 import Icon from 'components/Icon';
 import Modal from 'components/Modal';
 import NetworkSelector from 'components/NetworkSelector';
-import ProfileSignin from 'components/ProfileSignin';
 import Text from 'components/Text';
 
 import useAppSelector from 'hooks/useAppSelector';
 
-import headerNavClasses from './HeaderNav.module.scss';
-
-const LogoDesktopComponent = ui.logo.desktop ? Logos[ui.logo.desktop] : null;
-const LogoMobileComponent = ui.logo.mobile ? Logos[ui.logo.mobile] : null;
+import styles from './HeaderNav.module.scss';
 
 const headerNavMenu = Object.values(pages)
   .filter(page => page.enabled && page.navigation)
@@ -43,7 +38,7 @@ function HeaderNavModal({
   return (
     <>
       <Button
-        className={headerNavClasses.menu}
+        className={styles.menu}
         size="xs"
         variant="ghost"
         onClick={() => setShow(true)}
@@ -55,32 +50,32 @@ function HeaderNavModal({
         fullScreen
         fullWidth
         className={{
-          dialog: headerNavClasses.dialog
+          dialog: styles.dialog
         }}
       >
-        <header className={headerNavClasses.header}>
+        <header className={styles.header}>
           <Button
             size="xs"
             variant="ghost"
             onClick={handleHide}
-            className={headerNavClasses.hide}
+            className={styles.hide}
           >
             <Icon name="Cross" size="lg" title="Close Menu" />
           </Button>
         </header>
         {children(handleHide)}
-        <footer className={headerNavClasses.footer}>
+        <footer className={styles.footer}>
           {ui.layout.header.communityUrls.enabled && !isEmpty(community) ? (
             <div>
               <Text
                 color="gray"
                 scale="tiny-uppercase"
                 fontWeight="bold"
-                className={headerNavClasses.title}
+                className={styles.title}
               >
                 Join our community
               </Text>
-              <ul className={headerNavClasses.socials}>
+              <ul className={styles.socials}>
                 {community.map(social => (
                   <li key={social.name}>
                     <Text
@@ -89,12 +84,12 @@ function HeaderNavModal({
                       href={social.href}
                       target="_blank"
                       rel="noopener noreferrer"
-                      className={headerNavClasses.social}
+                      className={styles.social}
                     >
                       <Icon
                         title={social.name}
                         name={social.name}
-                        className={headerNavClasses.icon}
+                        className={styles.icon}
                       />
                     </Text>
                   </li>
@@ -105,7 +100,7 @@ function HeaderNavModal({
           <Feature name="regular">
             <CreateMarket
               fullwidth
-              className={headerNavClasses.createMarket}
+              className={styles.createMarket}
               onCreateClick={handleHide}
             />
           </Feature>
@@ -121,34 +116,28 @@ function HeaderNavMenu({
   onMenuItemClick?(): void;
 }>) {
   return (
-    <ul className={headerNavClasses.list}>
-      {headerNavMenu.map(page => (
-        <li key={page.name} className={headerNavClasses.item}>
-          <NavLink
-            to={page.pathname}
-            className={headerNavClasses.link}
-            activeClassName={headerNavClasses.active}
-            onClick={onMenuItemClick}
-            isActive={(_, location) =>
-              !!location.pathname.match(getPathname(page.pathname))
-            }
-          >
-            {page.name}
-          </NavLink>
-        </li>
-      ))}
-      {ui.layout.navbar.items.map(item => (
-        <li key={item.title} className={headerNavClasses.item}>
-          <a
-            target="_blank"
-            rel="noreferrer"
-            className={cn(headerNavClasses.link, headerNavClasses.customItem)}
-            href={item.href}
-          >
-            {item.title}
-          </a>
-        </li>
-      ))}
+    <ul className={styles.list}>
+      <li className={styles.item}>
+        <NavLink
+          to="/"
+          className={styles.link}
+          activeClassName={styles.active}
+          onClick={onMenuItemClick}
+        >
+          Communities
+        </NavLink>
+      </li>
+      <li className={styles.item}>
+        <NavLink
+          to="/"
+          className={styles.link}
+          activeClassName={styles.active}
+          onClick={onMenuItemClick}
+        >
+          Contests
+        </NavLink>
+      </li>
+
       {children}
     </ul>
   );
@@ -165,7 +154,7 @@ function HeaderNavMenuModal() {
             (ui.layout.header.helpUrl || ui.layout.onboarding) && (
               <>
                 {ui.layout.header.helpUrl && (
-                  <li className={headerNavClasses.item}>
+                  <li className={styles.item}>
                     <HelpButton
                       $outline
                       $fullWidth
@@ -175,7 +164,7 @@ function HeaderNavMenuModal() {
                   </li>
                 )}
                 {!!ui.layout.onboarding && (
-                  <li className={headerNavClasses.item}>
+                  <li className={styles.item}>
                     <HowToPlayButton $outline $fullWidth />
                   </li>
                 )}
@@ -195,46 +184,21 @@ export default function HeaderNav() {
     !!headerNavMenu.length &&
     !!ui.layout.navbar.items.length;
 
-  const headerLogo = () => {
-    if (LogoDesktopComponent && !LogoMobileComponent) {
-      return <LogoDesktopComponent />;
-    }
-
-    if (theme.device.isDesktop && LogoDesktopComponent) {
-      return <LogoDesktopComponent />;
-    }
-
-    if (!theme.device.isDesktop && LogoMobileComponent) {
-      return <LogoMobileComponent />;
-    }
-
-    return (
-      <>
-        <Logos.PolkamarketsLogo />
-        <V2Badge className={headerNavClasses.logosBadge} />
-      </>
-    );
-  };
-
   return (
-    <nav className={headerNavClasses.root}>
+    <nav className={styles.root}>
       {showLeftMenu && <HeaderNavMenuModal />}
       <Link
         to="/"
         aria-label="Homepage"
-        className={cn(headerNavClasses.logos, {
-          [headerNavClasses.logosGutter]: showLeftMenu
+        className={cn(styles.logos, {
+          [styles.logosGutter]: showLeftMenu
         })}
       >
-        {headerLogo()}
+        <Logos.PolkamarketsLogo />
       </Link>
       {theme.device.isTv && <HeaderNavMenu />}
       {!theme.device.isDesktop && ui.layout.header.networkSelector.enabled && (
-        <NetworkSelector
-          size="sm"
-          responsive
-          className={headerNavClasses.network}
-        />
+        <NetworkSelector size="sm" responsive className={styles.network} />
       )}
       {!theme.device.isDesktop && features.fantasy.enabled && !isLoggedIn && (
         <ProfileSignin variant="normal" color="primary" size="xs">
