--- src/components/ProfileSignout/ProfileSignout.module.scss
@@ -1,46 +1,0 @@
-.root {
-  display: flex;
-  align-items: center;
-  gap: 12px;
-  flex-direction: row-reverse;
-  flex: 1 0 auto;
-
-  @media (min-width: 1024px) {
-    flex-direction: unset;
-    margin-right: unset;
-  }
-}
-
-.signout,
-.transactions {
-  padding: var(--size-12) !important;
-  margin-left: auto;
-  color: var(--color-text-primary) !important;
-
-  @media (min-width: 1024px) {
-    margin-left: unset;
-  }
-
-  &Icon {
-    fill: var(--color-text-secondary) !important;
-  }
-}
-
-.username {
-  color: var(--color-text-primary);
-}
-
-.claim {
-  background-color: #5d55fa;
-  border-radius: 4px;
-  height: 24px;
-  padding-left: 8px;
-  padding-right: 8px;
-  border: 0px;
-  font-size: 14px;
-  text-transform: uppercase;
-  font-weight: 500;
-  margin-top: 2px;
-  color: var(--color-text-primary) !important;
-  cursor: pointer;
-}
