--- src/components/TransactionsButton/TransactionsButton.module.scss
@@ -1,7 +1,0 @@
-.root {
-  color: var(--color-text-primary) !important;
-}
-
-.icon {
-  fill: transparent !important;
-}
